refactor: (#808) Port 분리
