+++ dms-core/src/main/kotlin/team/aliens/dms/domain/vote/spi/CommandVotingOptionPort.kt
@@ -1,0 +1,11 @@
+package team.aliens.dms.domain.vote.spi
+
+import team.aliens.dms.domain.vote.model.VotingOption
+import java.util.UUID
+
+interface CommandVotingOptionPort {
+
+    fun saveVotingOption(votingOption: VotingOption): VotingOption
+
+    fun deleteVotingOptionByVotingOptionId(votingOptionId: UUID)
+}
