+++ dms-core/src/main/kotlin/team/aliens/dms/domain/vote/spi/VotingOptionPort.kt
@@ -1,0 +1,5 @@
+package team.aliens.dms.domain.vote.spi
+
+interface VotingOptionPort :
+    QueryVotingOptionPort,
+    CommandVotingOptionPort
