+++ dms-core/src/main/kotlin/team/aliens/dms/domain/vote/spi/VotingTopicPort.kt
@@ -1,0 +1,5 @@
+package team.aliens.dms.domain.vote.spi
+
+interface VotingTopicPort :
+    QueryVotingTopicPort,
+    CommandVotingTopicPort
