+++ dms-persistence/src/main/kotlin/team/aliens/dms/persistence/vote/VotePersistenceAdapter.kt
@@ -4,8 +4,6 @@ import com.querydsl.jpa.impl.JPAQueryFactory
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
 import team.aliens.dms.domain.vote.model.Vote
-import team.aliens.dms.domain.vote.model.VotingOption
-import team.aliens.dms.domain.vote.model.VotingTopic
 import team.aliens.dms.domain.vote.spi.VotePort
 import team.aliens.dms.domain.vote.spi.vo.OptionVotingResultVO
 import team.aliens.dms.domain.vote.spi.vo.StudentVotingResultVO
@@ -14,11 +12,7 @@ import team.aliens.dms.persistence.vote.entity.QVoteJpaEntity.voteJpaEntity
 import team.aliens.dms.persistence.vote.entity.QVotingOptionJpaEntity.votingOptionJpaEntity
 import team.aliens.dms.persistence.vote.entity.QVotingTopicJpaEntity.votingTopicJpaEntity
 import team.aliens.dms.persistence.vote.mapper.VoteMapper
-import team.aliens.dms.persistence.vote.mapper.VotingOptionMapper
-import team.aliens.dms.persistence.vote.mapper.VotingTopicMapper
 import team.aliens.dms.persistence.vote.repository.VoteJpaRepository
-import team.aliens.dms.persistence.vote.repository.VotingOptionJpaRepository
-import team.aliens.dms.persistence.vote.repository.VotingTopicJpaRepository
 import team.aliens.dms.persistence.vote.repository.vo.QQueryOptionVotingResultVO
 import team.aliens.dms.persistence.vote.repository.vo.QQueryStudentVotingResultVO
 import java.util.UUID
@@ -26,67 +20,19 @@ import java.util.UUID
 @Component
 class VotePersistenceAdapter(
     private val voteMapper: VoteMapper,
-    private val votingTopicMapper: VotingTopicMapper,
-    private val votingOptionMapper: VotingOptionMapper,
     private val voteJpaRepository: VoteJpaRepository,
-    private val votingTopicJpaRepository: VotingTopicJpaRepository,
-    private val votingOptionJpaRepository: VotingOptionJpaRepository,
     private val queryFactory: JPAQueryFactory
 ) : VotePort {
-
-    override fun saveVotingTopic(votingTopic: VotingTopic) = votingTopicMapper.toDomain(
-        votingTopicJpaRepository.save(
-            votingTopicMapper.toEntity(votingTopic)
-        )
-    )!!
-
-    override fun deleteVotingTopicById(votingTopicId: UUID) {
-        votingTopicJpaRepository.deleteById(votingTopicId)
-    }
-
-    override fun queryVotingTopicById(votingTopicId: UUID) = votingTopicMapper.toDomain(
-        votingTopicJpaRepository.findByIdOrNull(votingTopicId)
-    )
-
-    override fun queryAllVotingTopic() = votingTopicJpaRepository.findAll().map {
-        votingTopicMapper.toDomain(it)!!
-    }
 
     override fun saveVote(vote: Vote): Vote = voteMapper.toDomain(
         voteJpaRepository.save(voteMapper.toEntity(vote))
     )!!
 
-    override fun saveVotingOption(votingOption: VotingOption): VotingOption = votingOptionMapper.toDomain(
-        votingOptionJpaRepository.save(votingOptionMapper.toEntity(votingOption))
-    )!!
-    override fun deleteVotingTopicByVotingTopicId(
-        votingTopicId: UUID
-    ) = votingTopicJpaRepository.deleteById(votingTopicId)
-
-    override fun deleteVotingOptionByVotingOptionId(votingOptionId: UUID) = votingOptionJpaRepository.deleteById(votingOptionId)
     override fun deleteVoteById(voteId: UUID) = voteJpaRepository.deleteById(voteId)
-
-    override fun queryVotingOptionById(votingOptionId: UUID): VotingOption? = votingOptionMapper.toDomain(
-        votingOptionJpaRepository.findByIdOrNull(votingOptionId)
-    )
 
     override fun queryVoteById(voteId: UUID): Vote? = voteMapper.toDomain(
         voteJpaRepository.findByIdOrNull(voteId)
     )
-
-    override fun queryVotingOptionsByVotingTopicId(votingTopicId: UUID): List<VotingOption> {
-        return queryFactory
-            .selectFrom(votingOptionJpaEntity)
-            .where(votingOptionJpaEntity.votingTopic.id.eq(votingTopicId))
-            .fetch()
-            .map { entity ->
-                VotingOption(
-                    id = entity.id!!,
-                    votingTopicId = entity.votingTopic!!.id!!,
-                    optionName = entity.optionName
-                )
-            }
-    }
 
     override fun queryStudentVotingByVotingTopicIdAndGrade(votingTopicId: UUID, grade: Int): List<StudentVotingResultVO> {
         return queryFactory
@@ -125,15 +71,6 @@ class VotePersistenceAdapter(
             .orderBy(voteJpaEntity.id.count().intValue().desc())
             .fetch()
     }
-
-    override fun existVotingTopicByName(votingTopicName: String): Boolean =
-        votingTopicJpaRepository.existsByTopicName(votingTopicName)
-
-    override fun existVotingTopicById(votingTopicId: UUID): Boolean =
-        votingTopicJpaRepository.existsById(votingTopicId)
-
-    override fun existVotingOptionById(votingOptionId: UUID): Boolean =
-        votingOptionJpaRepository.existsById(votingOptionId)
 
     override fun queryVoteByStudentId(studentId: UUID): List<Vote> =
         voteJpaRepository.findByStudentId(studentId).map {
